Moved comment model tests into their own file
